Vinkel funker (ich)
lite problem ved origo, fordi når x er 0 og støy gjer at y = 1 så hopper vinkelen frå 0 til 1, men tenker vi kan legge til eit krav om at x eller y er over ein viss treshold for at man skal lese vinkelen så er det good

+++ ATmega162/ATmega162/Utils.h
@@ -22,7 +22,7 @@
 #define loopUntilBitIsSet(reg, bit) while(!testBit(reg, bit))
 #define loopUntilBitIsClear(reg, bit) while(testBit(reg, bit))
 
-float claculateJoystickAngel(int16_t x, int16_t y);
+int claculateJoystickAngel(int16_t x, int16_t y);
 
 
 #endif /* UTILS_H_ */
